Replaced error status pointers with assert() in Chebyshev polynomial functions.
Error handling in C is not always very graceful. We can't use exceptions
like we can in C++. One way to deal with errors is to include "success
pointers" as arguments to functions which can cause errors. However, I
found a better philosophy while prowling StackOverflow which said
something along the lines of the following:

If the only possible error is a programming error, use assert().

The errors in the Chebyshev polynomials occur if the index "n" is less
than 0. But that's a programming error. So for those I use assert() now.
It uses less code and works just as well.